+++ src/LN/clsCarrera.java
@@ -7,7 +7,7 @@ public class clsCarrera extends clsActividad
 	private double km;
 	private String ritmo;	
 
-	public clsCarrera (String fecha, String duracion, int calorias, double km, String ritmo) 
+	public clsCarrera (String fecha, String duracion, double calorias, double km, String ritmo) 
 	{
 		super(fecha, duracion, calorias);
 		this.km = km;
